stereo: profile switches interpolate the effects and focus zone too
Fade/Haze/Blur become floats in the GPU state (their consumers already
divide by 8.0f), and the profile lerp now glides them - plus the focus
zone edges - from the current values to the target's, alongside the
depths. A mid-transition re-switch still starts from wherever the
previous glide stopped. Only Edge Cleanup's mode snaps (a mode cannot
interpolate; its crop width already follows the lerped depths).

+++ source/3dsgpu.h
@@ -306,9 +306,11 @@ typedef struct
     // zone's back edge only (distance cues). Relative model: the layer
     // farthest outside receives the full gauge strength
     // (stereoMaxExcess for blur, stereoMaxBackExcess for fade/haze).
-    int                         stereoFade;
-    int                         stereoHaze;
-    int                         stereoBlur;
+    // float so profile switches can interpolate them (they lerp along
+    // with the depths; only Edge Cleanup's mode snaps)
+    float                       stereoFade;
+    float                       stereoHaze;
+    float                       stereoBlur;
     float                       stereoFocusBack;
     float                       stereoFocusFront;
     float                       stereoMaxExcess;
